pyproject.toml
[tool.poetry]
name = "quotationbuilder"
version = "0.1.0"
description = ""
authors = ["ddiederen <dirkdiederen@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
psycopg2-binary = "^2.9.5"
pygraphviz = "^1.10"
django-extensions = "^3.2.1"
django-crispy-forms = "^2.0"
djangorestframework = "^3.14.0"
Pillow = "^9.5.0"
django-filter = "^23.2"
PyYAML = "^6.0"
uritemplate = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"